third_party/WebKit/LayoutTests/external/wpt/webaudio/the-audio-api/the-analysernode-interface/test-analysernode-expected.txt

This is a testharness.js-based test.
FAIL Test AnalyserNode API assert_equals: analyser node has 2 input channels by default expected 2 but got 1
FAIL Test AnalyserNode's ctor API assert_equals: analyser node has 2 input channels by default expected 2 but got 1
Harness: the test ran to completion.
